Exception update

+++ pokemon/src/main/java/com/pokemon/service/PokemonServiceImp.java
@@ -17,18 +17,26 @@ public class PokemonServiceImp implements PokemonService{
 	@Override
 	@Transactional(readOnly = false)
 	public Pokemon save(Pokemon pokemon) {
+		if(pokemon.getIdPokedex() == null)
+			throw new PokemonNotFoundException("Nope");
+			
 		return pokemonDao.save(pokemon);
 	}
 
 	@Override
 	@Transactional(readOnly = true)
 	public List<Pokemon> findAll(){
+		if(((List<Pokemon>) pokemonDao.findAll()).isEmpty())
+			throw new PokemonNotFoundException("There is no Pokemon object");
 		return (List<Pokemon>) pokemonDao.findAll();
 	}
 
 	@Override
 	@Transactional(readOnly = false)
 	public void delete(Integer id) {
+		if(pokemonDao.findById(id).isEmpty())
+			throw new PokemonNotFoundException("Requested Pokemon does not exist");
+
 		pokemonDao.deleteById(id);
 	}
